requirements.txt
anthropic>=0.84.0
openai>=2.26.0
click>=8.3.1
rich>=14.3.0
python-dotenv>=1.2.0
